[question + proposition] Admin Dashboard pagination improvement
I am puzzled about this file, since in my version I have additional <form> block with selector (10/25/50/100, default 10), and cannot locate in history when and why those changes occured and why is this selector removed. I was going to propose improvement with above proposed changes in the code.

This is a proposition for Admin Dashboard pagination improvement, change current pagination settings from default 10 and set 10,25,50,100 to a larger set of default 100 and set options for 100,250,500,1000.